+++ packages/icon/templates/index.js
