+++ packages/mobx/package.json
@@ -2,6 +2,7 @@
   "name": "@learn-react/mobx",
   "version": "1.0.0",
   "license": "UNLICENSED",
+  "type": "module",
   "keywords": [
     "mobx",
     "mobx-react"
@@ -10,9 +11,9 @@
     "@learn-react/core": "1.0.0"
   },
   "scripts": {
-    "develop": "node ./bin/vite.mjs",
+    "develop": "node ./bin/vite",
     "type-check": "run --top-level tsc --noEmit",
     "start": "run-p \"develop\" \"type-check --watch\"",
-    "build": "yarn type-check && node ./bin/vite.mjs -m build"
+    "build": "yarn type-check && node ./bin/vite -m build"
   }
 }
