+++ packages/statement/package.json
@@ -2,11 +2,12 @@
   "name": "@learn-react/statement",
   "version": "1.0.0",
   "license": "UNLICENSED",
+  "type": "module",
   "dependencies": {
     "@learn-react/core": "1.0.0"
   },
   "scripts": {
-    "develop": "node ./bin/vite.mjs",
+    "develop": "node ./bin/vite",
     "type-check": "run --top-level tsc --noEmit",
     "start": "run-p \"type-check --watch\" \"develop\"",
     "build": "yarn type-check && node ./bin/vite -m build"
